refactor(skills): split skill helper modules
