+++ backend/onyx/server/features/skill/mutation_helpers.py
